Implemented TokenBucket Algorithm and modified existing FixedBucket Algorithm

+++ src/main/java/com/ratelimiter/rate_limiter_service/dto/CheckResponse.java
@@ -11,7 +11,7 @@ public class CheckResponse {
 
     boolean allowed;
 
-    int remaining;
+    double remaining;
 
     long retryAfter;
